feature(today-avarage-component): creating files
