+++ challenge-catar Watch App/Components/TextView/TextView.swift
@@ -6,3 +6,18 @@
 //
 
 import SwiftUI
+
+struct TextView: View {
+    
+    let text: String
+    let color: Color
+    let type: TextType
+    
+    var body: some View {
+        Text(text)
+            .font(.system(size: type.textStyle.size,
+                          weight: type.textStyle.weight,
+                          design: .rounded))
+            .foregroundColor(color)
+    }
+}
