+++ challenge-catar Watch App/Models/CardValues.swift
@@ -1,0 +1,13 @@
+//
+//  CardValues.swift
+//  challenge-catar Watch App
+//
+//  Created by alexdamascena on 27/11/22.
+//
+
+import Foundation
+
+struct CardValues {
+    let leftSideContent: String
+    let rightSideContent: String
+}
