+++ challenge-catar Watch App/Models/TextStyle.swift
@@ -1,0 +1,13 @@
+//
+//  TextStyle.swift
+//  challenge-catar Watch App
+//
+//  Created by alexdamascena on 27/11/22.
+//
+
+import Foundation
+
+struct TextStyle {
+    let size: CGFloat
+    let weight: Font.Weight
+}
